refactor: consolidate types and relocate utils under components
