+++ src/components/list-item-medium-and-venue/ListItemMediumAndVenue.tsx
@@ -8,9 +8,7 @@ export function ListItemMediumAndVenue({
   const value = medium && venue ? `${medium} at ${venue}` : (medium ?? venue);
 
   if (!value) {
-    {
-      /* pre-2012 viewings don't have venue or medium */
-    }
+    // pre-2012 viewings don't have venue or medium
     return undefined;
   }
