+++ src/components/list-item-title/ListItemTitle.tsx
@@ -1,20 +1,11 @@
-/**
- * Renders a movie title with year, optionally as a link to the review.
- * @param props - Component props
- * @param props.className - Additional CSS classes
- * @param props.slug - URL slug for linking to the review (if reviewed)
- * @param props.title - The movie title
- * @param props.year - The movie release year
- * @returns Title with year, linked if slug is provided
- */
 export function ListItemTitle({
   className = "",
-  slug,
+  reviewSlug,
   title,
   year,
 }: {
   className?: string;
-  slug?: string;
+  reviewSlug?: string;
   title: string;
   year: string;
 }): React.JSX.Element {
@@ -24,25 +15,21 @@ export function ListItemTitle({
     </span>
   );
 
-  if (slug) {
-    return (
-      <a
-        className={`
-          text-base/5 font-semibold text-default transition-all duration-500
-          after:absolute after:top-0 after:left-0 after:z-sticky after:size-full
-          after:opacity-0
-          hover:text-accent
-        `}
-        href={`/reviews/${slug}/`}
-      >
-        {title}
-        &#x202F;&#x202F;
-        {yearBox}
-      </a>
-    );
-  }
-
-  return (
+  return reviewSlug ? (
+    <a
+      className={`
+        text-base/5 font-semibold text-default transition-all duration-500
+        after:absolute after:top-0 after:left-0 after:z-sticky after:size-full
+        after:opacity-0
+        hover:text-accent
+      `}
+      href={`/reviews/${reviewSlug}/`}
+    >
+      {title}
+      &#x202F;&#x202F;
+      {yearBox}
+    </a>
+  ) : (
     <span
       className={`
         ${className}
