+++ src/features/reviews/ReviewsListItem.tsx
@@ -1,6 +1,3 @@
-import type { PosterImageProps } from "~/assets/posters";
-import type { GradeText, GradeValue } from "~/utils/grades";
-
 import { ListItemDetails } from "~/components/list-item-details/ListItemDetails";
 import { ListItemGenres } from "~/components/list-item-genres/ListItemGenres";
 import { ListItemGrade } from "~/components/list-item-grade/ListItemGrade";
@@ -8,28 +5,8 @@ import { ListItemReviewDate } from "~/components/list-item-review-date/ListItemR
 import { ListItemTitle } from "~/components/list-item-title/ListItemTitle";
 import { PosterListItem } from "~/components/poster-list/PosterListItem";
 
-export type ReviewsValue = {
-  genres: string[];
-  grade: GradeText;
-  gradeValue: GradeValue;
-  imdbId: string;
-  posterImageProps: PosterImageProps;
-  releaseSequence: number;
-  releaseYear: string;
-  reviewDisplayDate: string;
-  reviewSequence: string;
-  reviewYear: string;
-  slug: string;
-  sortTitle: string;
-  title: string;
-};
+import type { ReviewsValue } from "./Reviews";
 
-/**
- * List item component for displaying a single review.
- * @param props - Component props
- * @param props.value - Review data to display
- * @returns Review list item with poster, title, grade, and details
- */
 export function ReviewsListItem({
   value,
 }: {
@@ -39,7 +16,7 @@ export function ReviewsListItem({
     <PosterListItem posterImageProps={value.posterImageProps}>
       <ListItemDetails>
         <ListItemTitle
-          slug={value.slug}
+          reviewSlug={value.slug}
           title={value.title}
           year={value.releaseYear}
         />
