+++ src/features/stats/MostWatchedMovies.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from "astro:content";
 
-import ListItemTitle from "~/components/list-item-title/ListItemTitle.astro";
+import { ListItemTitle } from "~/components/list-item-title/ListItemTitle";
 import PosterList from "~/components/poster-list/PosterList.astro";
 import PosterListItem from "~/components/poster-list/PosterListItem.astro";
