+++ src/features/stats/MostWatchedPeople.astro
@@ -3,7 +3,7 @@ import type { MostWatchedPerson } from "~/collections/stats";
 
 import { ListItemMediumAndVenue } from "~/components/list-item-medium-and-venue/ListItemMediumAndVenue";
 import { ListItemReviewDate } from "~/components/list-item-review-date/ListItemReviewDate";
-import ListItemTitle from "~/components/list-item-title/ListItemTitle.astro";
+import { ListItemTitle } from "~/components/list-item-title/ListItemTitle";
 import PosterList from "~/components/poster-list/PosterList.astro";
 import PosterListItem from "~/components/poster-list/PosterListItem.astro";
 import { toDisplayDate } from "~/components/utils/toDisplayDate";
